# Optimize SearchSidebar by simplifying loading logic