@mountiny 2025-07-22
@allroundexperts can you please test the original regression and then test if the loading bar is not showing when you are completely offline